fix type conversion warnings

+++ Adjust/ADJUtil.h
@@ -122,6 +122,6 @@ typedef void (^isInactiveInjected)(BOOL);
 + (NSString *)decimalToBinaryString:(NSUInteger)decInt;
 
 + (NSString *)enforceParameterLength:(NSString *)parameter
-                       withMaxlength:(int)maxLength;
+                       withMaxlength:(NSUInteger)maxLength;
 
 @end
